04_environment/azure-ai-python/requirements.txt
azure-ai-projects ==1.0.0b11
azure-identity == 1.23.0
azure-storage-blob == 12.25.1
semantic-kernel == 1.34.0 
pandas == 2.3.1
ipykernel == 6.29.5